# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knotline internal API",
    "version": "0.1.0",
    "description": "Generated from the repository-owned transport contract registry."
  },
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Local development"
    }
  ],
  "paths": {
    "/edge/v1/auth/magic-links": {
      "post": {
        "operationId": "requestMagicLink",
        "summary": "Request a non-enumerating one-time email sign-in link",
        "tags": [
          "Authentication"
        ],
        "x-knotline-exposure": "public_anonymous",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "http://json-schema.org/draft-07/schema#",
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "intent": {
                    "default": "login",
                    "type": "string",
                    "enum": [
                      "login",
                      "step_up"
                    ]
                  },
                  "returnTargetId": {
                    "default": "workflows",
                    "type": "string",
                    "maxLength": 40
                  }
                },
                "required": [
                  "email",
                  "intent",
                  "returnTargetId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful response",
            "headers": {
              "Knotline-Request-Id": {
                "description": "Stable request correlation identifier",
                "schema": {
                  "type": "string"
                }
              },
              "traceparent": {
                "description": "W3C distributed tracing context",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "accepted"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Typed error response",
            "headers": {
              "Knotline-Request-Id": {
                "description": "Stable request correlation identifier",
                "schema": {
                  "type": "string"
                }
              },
              "traceparent": {
                "description": "W3C distributed tracing context",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 500
                        },
                        "requestId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "details": {
                          "type": "object",
                          "propertyNames": {
                            "type": "string"
                          },
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Typed error response",
            "headers": {
              "Knotline-Request-Id": {
                "description": "Stable request correlation identifier",
                "schema": {
                  "type": "string"
                }
              },
              "traceparent": {
                "description": "W3C distributed tracing context",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 500
                        },
                        "requestId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "details": {
                          "type": "object",
                          "propertyNames": {
                            "type": "string"
                          },
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}